refactor: 카테고리 삭제 전략 패턴 repository 주입으로 변경
