+++ src/main/java/shop/wannab/book_service/category/repository/CategoryQueryRepositoryImpl.java
@@ -7,7 +7,9 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
+import shop.wannab.book_service.book.entity.QBookCategory;
 import shop.wannab.book_service.category.dto.response.CategoryResponse;
+import shop.wannab.book_service.category.entity.Category;
 import shop.wannab.book_service.category.entity.QCategory;
 
 @RequiredArgsConstructor
@@ -15,10 +17,11 @@ public class CategoryQueryRepositoryImpl implements CategoryQueryRepository{
 
     private final JPAQueryFactory queryFactory;
 
+    private final QCategory category = QCategory.category;
+    private final QBookCategory bookCategory = QBookCategory.bookCategory;
+
     @Override
     public List<CategoryResponse> findParentCategories() {
-        QCategory category = QCategory.category;
-
         return queryFactory
                 .select(Projections.constructor(CategoryResponse.class,
                         category.id,
@@ -30,8 +33,6 @@ public class CategoryQueryRepositoryImpl implements CategoryQueryRepository{
 
     @Override
     public Page<CategoryResponse> findParentCategories(Pageable pageable) {
-        QCategory category = QCategory.category;
-
         List<CategoryResponse> content = queryFactory
                 .select(Projections.constructor(CategoryResponse.class,
                         category.id,
@@ -53,8 +54,6 @@ public class CategoryQueryRepositoryImpl implements CategoryQueryRepository{
 
     @Override
     public Page<CategoryResponse> findChildCategoriesByParentId(Long parentId, Pageable pageable) {
-        QCategory category = QCategory.category;
-
         List<CategoryResponse> content = queryFactory
                 .select(Projections.constructor(CategoryResponse.class,
                         category.id,
@@ -72,5 +71,38 @@ public class CategoryQueryRepositoryImpl implements CategoryQueryRepository{
                 .fetchOne();
 
         return new PageImpl<>(content, pageable, total != null ? total : 0);
+    }
+
+    @Override
+    public void deleteCategoryWithBookCategories(Category targetCategory) {
+        Long categoryId = targetCategory.getId();
+
+        List<Long> childCategoryIds = queryFactory
+                .select(category.id)
+                .from(category)
+                .where(category.parent.id.eq(categoryId))
+                .fetch();
+
+        if (!childCategoryIds.isEmpty()) {
+            queryFactory
+                    .delete(bookCategory)
+                    .where(bookCategory.category.id.in(childCategoryIds))
+                    .execute();
+
+            queryFactory
+                    .delete(category)
+                    .where(category.id.in(childCategoryIds))
+                    .execute();
+        }
+
+        queryFactory
+                .delete(bookCategory)
+                .where(bookCategory.category.id.eq(categoryId))
+                .execute();
+
+        queryFactory
+                .delete(category)
+                .where(category.id.eq(categoryId))
+                .execute();
     }
 }
