+++ src/main/java/shop/wannab/book_service/category/service/strategy/PhysicalCategoryDeleteStrategy.java
@@ -7,19 +7,19 @@ import shop.wannab.book_service.category.entity.Category;
 import shop.wannab.book_service.category.repository.CategoryRepository;
 
 /**
- * 해당 카테고리 데이터를 논리적으로 삭제하는 전략
+ * 해당 카테고리 데이터를 물리적으로 삭제하는 전략
+ * 추후 논리적 삭제로 변경 예정
  */
 @Component
 @RequiredArgsConstructor
-public class SoftDeleteStrategy implements CategoryDeleteStrategy {
+public class PhysicalCategoryDeleteStrategy implements CategoryDeleteStrategy {
 
     private final CategoryRepository categoryRepository;
 
     @Override
     @Transactional
-    public void delete(Category category) {
-        category.delete();
-        categoryRepository.save(category);
+    public void delete(Category targetCategory) {
+        categoryRepository.deleteCategoryWithBookCategories(targetCategory);
     }
 }
